Extract the stack trace sanitization logic out of `FeatureUsage` and into a separate class.
