+++ subprojects/logging/src/test/groovy/org/gradle/internal/featurelifecycle/StackTraceSanitizerTest.groovy
@@ -23,12 +23,12 @@ import static SimulatedSingleMessageLogger.DIRECT_CALL
 import static SimulatedSingleMessageLogger.INDIRECT_CALL
 import static SimulatedSingleMessageLogger.INDIRECT_CALL_2
 
-@Subject(FeatureUsage)
-class FeatureUsageTest extends Specification {
+@Subject(StackTraceSanitizer)
+class StackTraceSanitizerTest extends Specification {
 
     def "stack is evaluated correctly for #callLocationClass.simpleName and #expectedSummary. #expectedMethod. #usage."() {
         given:
-        def stack = FeatureUsage.calculateStack(usage.calledFrom, usage.exception.stackTrace)
+        def stack = new StackTraceSanitizer(usage.calledFrom).transform(usage.exception.stackTrace)
 
         expect:
         !stack.empty
